requirements.txt
# RAG assignment dependencies
sentence-transformers
faiss-cpu
flask
flask-cors
pyyaml
groq
networkx
numpy
